+++ ZFSin/zfs/include/sys/u8_textprep.h
@@ -54,14 +54,14 @@ extern "C" {
 #define	UCONV_IN_ACCEPT_BOM		0x0080
 #define	UCONV_OUT_EMIT_BOM		0x0100
 
-extern int uconv_u16tou32(const uint16_t *, uint32_t *, uint32_t *, uint32_t *,
+extern int uconv_u16tou32(const uint16_t *, size_t *, uint32_t *, size_t *,
 	int);
-extern int uconv_u16tou8(const uint16_t *, uint32_t *, uchar_t *, uint32_t *, int);
-extern int uconv_u32tou16(const uint32_t *, uint32_t *, uint16_t *, uint32_t *,
+extern int uconv_u16tou8(const uint16_t *, size_t *, uchar_t *, size_t *, int);
+extern int uconv_u32tou16(const uint32_t *, size_t *, uint16_t *, size_t *,
 	int);
-extern int uconv_u32tou8(const uint32_t *, uint32_t *, uchar_t *, uint32_t *, int);
-extern int uconv_u8tou16(const uchar_t *, uint32_t *, uint16_t *, uint32_t *, int);
-extern int uconv_u8tou32(const uchar_t *, uint32_t *, uint32_t *, uint32_t *, int);
+extern int uconv_u32tou8(const uint32_t *, size_t *, uchar_t *, size_t *, int);
+extern int uconv_u8tou16(const uchar_t *, size_t *, uint16_t *, size_t *, int);
+extern int uconv_u8tou32(const uchar_t *, size_t *, uint32_t *, size_t *, int);
 
 /*
  * UTF-8 text preparation functions and their macros.
@@ -106,9 +106,9 @@ extern int uconv_u8tou32(const uchar_t *, uint32_t *, uint32_t *, uint32_t *, in
 #define	U8_ILLEGAL_CHAR			(-1)
 #define	U8_OUT_OF_RANGE_CHAR		(-2)
 
-extern int u8_validate(char *, uint32_t, char **, int, int *);
-extern int u8_strcmp(const char *, const char *, uint32_t, int, uint32_t, int *);
-extern uint32_t u8_textprep_str(char *, uint32_t *, char *, uint32_t *, int, uint32_t,
+extern int u8_validate(char *, size_t, char **, int, int *);
+extern int u8_strcmp(const char *, const char *, size_t, int, size_t, int *);
+extern size_t u8_textprep_str(char *, size_t *, char *, size_t *, int, size_t,
 	int *);
 
 #ifdef	__cplusplus
