+++ ZFSin/zfs/module/icp/io/sha2_mod.c
@@ -178,7 +178,7 @@ static crypto_mac_ops_t sha2_mac_ops = {
 
 static int sha2_create_ctx_template(crypto_provider_handle_t,
     crypto_mechanism_t *, crypto_key_t *, crypto_spi_ctx_template_t *,
-    uint32_t *, crypto_req_handle_t);
+    size_t *, crypto_req_handle_t);
 static int sha2_free_context(crypto_ctx_t *);
 
 static crypto_ctx_ops_t sha2_ctx_ops = {
@@ -1579,7 +1579,7 @@ bail:
 static int
 sha2_create_ctx_template(crypto_provider_handle_t provider,
     crypto_mechanism_t *mechanism, crypto_key_t *key,
-    crypto_spi_ctx_template_t *ctx_template, uint32_t *ctx_template_size,
+    crypto_spi_ctx_template_t *ctx_template, size_t *ctx_template_size,
     crypto_req_handle_t req)
 {
 	sha2_hmac_ctx_t *sha2_hmac_ctx_tmpl;
