+++ ZFSin/zfs/module/zfs/spa_config.c
@@ -154,7 +154,7 @@ out:
 static void
 spa_config_write(spa_config_dirent_t *dp, nvlist_t *nvl)
 {
-	uint32_t buflen;
+	size_t buflen;
 	char *buf;
 	vnode_t *vp;
 	int oflags = FWRITE | FTRUNC | FCREAT | FOFFMAX;
